refactor: complete rewrite
I did not like the organization of the previous code, so I decided to
start from zero again. Feature-wise, this one supports filtering on top
of from everything the older version already supported. The UI changed
slightly because I changed the way the lists works. This time I tried to
make it more modular and with less functionality spread around many
functions in many files.